Fix up receiver declaration.
The intent this receiver listens to is used by the BT MAP implementation
to send out SMS it queued. As the manifest and the
SmsReceiverService.java were in disagreement about the action name, this
didn't work.
Also secure the receiver in the process, it can be used to send out SMS
without having the SEND_SMS permission otherwise.

Change-Id: I3ef0be58c641aacdce4b42ec9e6654087948977d