MMS: Disable Android 4.2 specific hack regarding keyboard behavior.
In android 4.2 the keyboard activity resizing events were delayed by about
a half a second and there was no fade animation. This caused some apps
to behave oddly when displaying the keyboard upon loading including
the MMS app which google had implemented a hack to workaround this glitch.
In android 4.3 and newer the old 4.0-4.1 behavior was restored, making this
hack unnecessary. Disabling the hack removes the 1 second delay from the
message view displaying that was necessary for the 4.2 behavior but no longer
needed.

Change-Id: Ia9ba1d9217e26508285c1aea70aff9788a19f95e